move event and property definitions to models, refactor models to import via component if user present
