+++ frontend/src/layout/navigation/MainNavigation.tsx
@@ -30,11 +30,11 @@ import {
 } from 'lib/components/icons'
 import { navigationLogic } from './navigationLogic'
 import { ToolbarModal } from '~/layout/ToolbarModal/ToolbarModal'
-import { dashboardsModel } from '~/models'
+import { dashboardsModel } from '~/models/dashboardsModel'
 import { DashboardType, HotKeys } from '~/types'
 import { userLogic } from 'scenes/userLogic'
 import { organizationLogic } from 'scenes/organizationLogic'
-import { canViewPlugins } from '../../scenes/plugins/access'
+import { canViewPlugins } from 'scenes/plugins/access'
 import { useGlobalKeyboardHotkeys, useKeyboardHotkeys } from 'lib/hooks/useKeyboardHotkeys'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { router } from 'kea-router'
