+++ frontend/src/scenes/actions/ActionEdit.js
@@ -9,7 +9,7 @@ import { Button, Col, Input, Row } from 'antd'
 import { InfoCircleOutlined, PlusOutlined, SaveOutlined, DeleteOutlined } from '@ant-design/icons'
 import { router } from 'kea-router'
 import { PageHeader } from 'lib/components/PageHeader'
-import { actionsModel } from '~/models'
+import { actionsModel } from '~/models/actionsModel'
 import { AsyncActionMappingNotice } from 'scenes/project/Settings/WebhookIntegration'
 import { preflightLogic } from 'scenes/PreflightCheck/logic'
 import dayjs from 'dayjs'
