+++ frontend/src/scenes/annotations/logic.ts
@@ -1,7 +1,7 @@
 import { kea } from 'kea'
 import api from 'lib/api'
 import { toParams, deleteWithUndo } from 'lib/utils'
-import { annotationsModel } from '~/models'
+import { annotationsModel } from '~/models/annotationsModel'
 import { annotationsTableLogicType } from './logicType'
 import { AnnotationType } from '~/types'
