+++ frontend/src/scenes/events/EventsVolumeTable.tsx
@@ -7,7 +7,7 @@ import { capitalizeFirstLetter, humanizeNumber } from 'lib/utils'
 import { preflightLogic } from 'scenes/PreflightCheck/logic'
 import { ColumnsType } from 'antd/lib/table'
 import { PropertyKeyInfo } from 'lib/components/PropertyKeyInfo'
-import { eventDefinitionsLogic } from './eventDefinitionsLogic'
+import { eventDefinitionsLogic } from '~/models/eventDefinitionsLogic'
 import { EventDefinition, PropertyDefinition } from '~/types'
 import { PageHeader } from 'lib/components/PageHeader'
 import { userLogic } from 'scenes/userLogic'
