+++ frontend/src/scenes/insights/BreakdownFilter/BreakdownFilter.js
@@ -4,8 +4,8 @@ import { useValues } from 'kea'
 import { propertyFilterLogic } from 'lib/components/PropertyFilters/propertyFilterLogic'
 import { PropertyKeyInfo } from 'lib/components/PropertyKeyInfo'
 import { SelectGradientOverflow } from 'lib/components/SelectGradientOverflow'
-import { propertyDefinitionsLogic } from 'scenes/events/propertyDefinitionsLogic'
-import { cohortsModel } from '~/models'
+import { propertyDefinitionsLogic } from '~/models/propertyDefinitionsLogic'
+import { cohortsModel } from '~/models/cohortsModel'
 import { ViewType } from '../insightLogic'
 
 const { TabPane } = Tabs
