+++ frontend/src/scenes/insights/TrendLegend/TrendLegend.tsx
@@ -5,7 +5,7 @@ import { IndexedTrendResult, trendsLogic } from 'scenes/trends/trendsLogic'
 import { PHCheckbox } from 'lib/components/PHCheckbox'
 import { getChartColors } from 'lib/colors'
 import { MATHS } from 'lib/constants'
-import { cohortsModel } from '~/models'
+import { cohortsModel } from '~/models/cohortsModel'
 import { CohortType } from '~/types'
 import { ColumnsType } from 'antd/lib/table'
 import { maybeAddCommasToInteger } from 'lib/utils'
